PR #25691: Fix an error testcase in debug_options_parsers_test
Imported from GitHub PR https://github.com/openxla/xla/pull/25691

The two tests are not independently parsing and testing like they seem to be. It is being parsed on the first call to PassFuelIsSet/ConsumeFuel (whichever executes first in random order of execution of tests). This then parses the XLA_FLAGS from the same test and in the other test, this is not parsed anymore. It seems to be passing because the second test uses the same XLA_FLAGS string. When this string is changed, the test starts failing.
Copybara import of the project:

--
bf286e724bfeb367349f24f34b418965198289c0 by Shraiysh Vaishay <svaishay@nvidia.com>:

Fix an error testcase in debug_options_parsers_test

The two tests are not independently parsing and testing like they seem to be.
It is being parsed on the first call to PassFuelIsSet/ConsumeFuel (whichever
executes first in random order of execution of tests). This then parses the
XLA_FLAGS from the same test and in the other test, this is not parsed anymore.
It seems to be passing because the second test uses the same XLA_FLAGS string.
When this string is changed, the test starts failing.

Merging this change closes #25691

FUTURE_COPYBARA_INTEGRATE_REVIEW=https://github.com/openxla/xla/pull/25691 from shraiysh:fix_erroneous_test_debug_options_parser bf286e724bfeb367349f24f34b418965198289c0
PiperOrigin-RevId: 752594816